requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
choicelib==0.1.5
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
pydantic==1.9.2
sniffio==1.2.0
typing_extensions==4.3.0
vbml==1.1.post1
vkbottle==4.3.9
vkbottle-types==5.131.146.11
watchfiles==0.16.1
yarl==1.8.1
pdfplumber~=0.9.0